requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-reset-migrations==0.4.0
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==13.12.0
forex-python==1.8
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==21.3
pycodestyle==2.8.0
pycountry==22.3.5
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
toposort==1.7
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wincertstore==0.2
